Api/myapi/requirements.txt
Django
djangorestframework
django-cors-headers
openai
